requirements.txt
# Feito com pip3 freeze
certifi==2021.10.8
charset-normalizer==2.0.12
exceptiongroup==1.2.1
idna==3.3
iniconfig==2.0.0
MouseInfo==0.1.3
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
PyAutoGUI==0.9.54  # Não sei se estamos usando mais
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
pytest==8.2.2
python3-xlib==0.15
pytweening==1.2.0
requests==2.27.1
tomli==2.0.1
urllib3==1.26.9

# Feito manualmente
# Pillow==10.3.0
# certifi==2021.10.8
# charset-normalizer==2.0.12
# idna==3.3
# requests==2.27.1
# urllib3==1.26.9
# PyAutoGUI==0.9.54
# pytest==8.2.2
